Corrected misaligned drills
Corrected misaligned motor mount hole (originally slightly too high and too far left); shifted FFC connector left slightly to align better with the cable bracket; widened cutout for POM adjuster slightly on the left; enlarged M2 mounting holes for cable bracket slightly. 

After receiving the first batch of boards, I noticed that I'd done math wrong on the motor mounting hole positions. (Did the math based on the edge, not the centre, because I'm stupid. I also made some slight tweaks to other items to get it to line up a little better with the original board's cutouts, but all in, it was actually pretty close!

One thing to note: The connectors that were previously SMD ZH connectors were in places where the mount for the sprite was totally flat, so you'll likely need to snip the legs down _very_ flush once the through-hole connectors are mounted. The single through-hole XH on the original board has a cutout in the mount behind it; you might consider filing out a similar groove behind the new through-holes in the other side.